Add fistures to import initial data
according to #9 is needed to import some initial Policy/Measure to
improve the usability of the tool.

This is a start, not completed yes.
Add fistures to import initial data

according to #9 is needed to import some initial Policy/Measure to
improve the usability of the tool.

This is a start, not completed yes.